pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "pygithub>=2.8.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
